yoko2redfish/requirements.txt
flask-restplus>=0.9.2
pyyaml>=3.10
dateutils>=0.6.6
